fix empty branch names getting returned by using uint64 id (#926)
The gatherBranches() function shifts a mask of type uint64 until it is 0, but
uses an id for accessing a map of type uint32. This results in empty branch
names getting returned from this function. It can be fixed by changing the type
of id to also be uint64.

+++ index/eval.go
@@ -498,7 +498,7 @@ func (d *indexData) gatherBranches(docID uint32, mt matchTree, known map[matchTr
 	}
 
 	var branches []string
-	id := uint32(1)
+	id := uint64(1)
 	branchNames := d.branchNames[d.repos[docID]]
 	for mask != 0 {
 		if mask&0x1 != 0 {
